+++ src/Type/Definition/InputObjectType.php
@@ -18,6 +18,7 @@ use GraphQL\Utils\Utils;
  * @phpstan-type InputObjectConfig array{
  *   name?: string|null,
  *   description?: string|null,
+ *   isOneOf?: bool|null,
  *   fields: iterable<FieldConfig>|callable(): iterable<FieldConfig>,
  *   parseValue?: ParseValueFn|null,
  *   astNode?: InputObjectTypeDefinitionNode|null,
@@ -27,6 +28,8 @@ use GraphQL\Utils\Utils;
 class InputObjectType extends Type implements InputType, NullableType, NamedType
 {
     use NamedTypeImplementation;
+
+    public bool $isOneOf;
 
     /**
      * Lazily initialized.
@@ -56,6 +59,7 @@ class InputObjectType extends Type implements InputType, NullableType, NamedType
     {
         $this->name = $config['name'] ?? $this->inferName();
         $this->description = $config['description'] ?? null;
+        $this->isOneOf = $config['isOneOf'] ?? false;
         // $this->fields is initialized lazily
         $this->parseValue = $config['parseValue'] ?? null;
         $this->astNode = $config['astNode'] ?? null;
@@ -94,6 +98,12 @@ class InputObjectType extends Type implements InputType, NullableType, NamedType
         }
 
         return isset($this->fields[$name]);
+    }
+
+    /** Returns true if this is a oneOf input object type. */
+    public function isOneOf(): bool
+    {
+        return $this->isOneOf;
     }
 
     /**
@@ -200,6 +210,39 @@ class InputObjectType extends Type implements InputType, NullableType, NamedType
 
         foreach ($resolvedFields as $field) {
             $field->assertValid($this);
+        }
+
+        // Additional validation for oneOf input objects
+        if ($this->isOneOf()) {
+            $this->validateOneOfConstraints($resolvedFields);
+        }
+    }
+
+    /**
+     * Validates that oneOf input object constraints are met.
+     *
+     * @param array<string, InputObjectField> $fields
+     *
+     * @throws InvariantViolation
+     */
+    private function validateOneOfConstraints(array $fields): void
+    {
+        if (count($fields) === 0) {
+            throw new InvariantViolation("OneOf input object type {$this->name} must define one or more fields.");
+        }
+
+        foreach ($fields as $fieldName => $field) {
+            $fieldType = $field->getType();
+
+            // OneOf fields must be nullable (not wrapped in NonNull)
+            if ($fieldType instanceof NonNull) {
+                throw new InvariantViolation("OneOf input object type {$this->name} field {$fieldName} must be nullable.");
+            }
+
+            // OneOf fields cannot have default values
+            if ($field->defaultValueExists()) {
+                throw new InvariantViolation("OneOf input object type {$this->name} field {$fieldName} cannot have a default value.");
+            }
         }
     }